Add Roact example
Roblox does not currently expose resolution scale to user code,
so to emulate its changing, this example has a mock provider
that constantly cycles the scaling factor.

An untested CoreProvider is also included in this example.

Eventually, this code should be turned into a library, but due
to the state of public-facing high DPI aware APIs in the engine,
it's unclear where it would live or how much it would help.